+++ src/lib.rs
@@ -1,9 +1,11 @@
 #![no_std]
 
 mod assert;
+mod current;
 mod helpers;
 mod resistance;
 mod voltage;
 
+pub use current::{Current, ExtF32 as CurrentExtF32, ExtU32 as CurrentExtU32};
 pub use resistance::{ExtF32 as ResistanceExtF32, ExtU32 as ResistanceExtU32, Resistance};
 pub use voltage::{ExtF32 as VoltageExtF32, ExtI32 as VoltageExtI32, Voltage};