+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - 2023-02-02
+## [0.1.0] - Unreleased
 
 ### Added
 
+- `Voltage` struct for storing voltage values (positive and negative)
+- `Voltage` extension methods for `i32` and `f32` values
 - `Resistance` struct for storing resistance values
 - `Resistance` extension methods for `u32` and `f32` values